BCF-2167: mv status and profile from `node` to `admin`
In order to consistently scope TOML configuration

- hides but does not rm the admin subcmds and adds a warning that points
to the updated location
- add subcmds to `admin` and reorganizes func locations to be in the same file
