+++ core/cmd/admin_commands.go
@@ -5,7 +5,13 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"path/filepath"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/manyminds/api2go/jsonapi"
 	"github.com/urfave/cli"
@@ -41,6 +47,29 @@ func initAdminSubCmds(client *Client) []cli.Command {
 			Name:   "logout",
 			Usage:  "Delete any local sessions",
 			Action: client.Logout,
+		},
+		{
+			Name:   "profile",
+			Usage:  "Collects profile metrics from the node.",
+			Action: client.Profile,
+			Flags: []cli.Flag{
+				cli.Uint64Flag{
+					Name:  "seconds, s",
+					Usage: "duration of profile capture",
+					Value: 8,
+				},
+				cli.StringFlag{
+					Name:  "output_dir, o",
+					Usage: "output directory of the captured profile",
+					Value: "/tmp/",
+				},
+			},
+		},
+		{
+			Name:   "status",
+			Usage:  "Displays the health of various services running inside the node.",
+			Action: client.Status,
+			Flags:  []cli.Flag{},
 		},
 		{
 			Name:  "users",
@@ -263,4 +292,121 @@ func (cli *Client) DeleteUser(c *cli.Context) (err error) {
 	}()
 
 	return cli.renderAPIResponse(response, &AdminUsersPresenter{}, "Successfully deleted API user")
+}
+
+// Status will display the health of various services
+func (cli *Client) Status(c *cli.Context) error {
+	resp, err := cli.HTTP.Get("/health?full=1", nil)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	defer func() {
+		if cerr := resp.Body.Close(); cerr != nil {
+			err = multierr.Append(err, cerr)
+		}
+	}()
+
+	return cli.renderAPIResponse(resp, &HealthCheckPresenters{})
+}
+
+// Profile will collect pprof metrics and store them in a folder.
+func (cli *Client) Profile(c *cli.Context) error {
+	seconds := c.Uint("seconds")
+	baseDir := c.String("output_dir")
+
+	genDir := filepath.Join(baseDir, fmt.Sprintf("debuginfo-%s", time.Now().Format(time.RFC3339)))
+
+	err := os.Mkdir(genDir, 0o755)
+	if err != nil {
+		return cli.errorOut(err)
+	}
+	var wgPprof sync.WaitGroup
+	vitals := []string{
+		"allocs",       // A sampling of all past memory allocations
+		"block",        // Stack traces that led to blocking on synchronization primitives
+		"cmdline",      // The command line invocation of the current program
+		"goroutine",    // Stack traces of all current goroutines
+		"heap",         // A sampling of memory allocations of live objects.
+		"mutex",        // Stack traces of holders of contended mutexes
+		"profile",      // CPU profile.
+		"threadcreate", // Stack traces that led to the creation of new OS threads
+		"trace",        // A trace of execution of the current program.
+	}
+	wgPprof.Add(len(vitals))
+	cli.Logger.Infof("Collecting profiles: %v", vitals)
+	cli.Logger.Infof("writing debug info to %s", genDir)
+
+	errs := make(chan error, len(vitals))
+	for _, vt := range vitals {
+		go func(vt string) {
+			defer wgPprof.Done()
+			uri := fmt.Sprintf("/v2/debug/pprof/%s?seconds=%d", vt, seconds)
+			resp, err := cli.HTTP.Get(uri)
+			if err != nil {
+				errs <- fmt.Errorf("error collecting %s: %w", vt, err)
+				return
+			}
+			defer func() {
+				if resp.Body != nil {
+					resp.Body.Close()
+				}
+			}()
+			if resp.StatusCode == http.StatusUnauthorized {
+				errs <- fmt.Errorf("error collecting %s: %w", vt, errUnauthorized)
+				return
+			}
+			if resp.StatusCode == http.StatusBadRequest {
+				// best effort to interpret the underlying problem
+				pprofVersion := resp.Header.Get("X-Go-Pprof")
+				if pprofVersion == "1" {
+					b, err := io.ReadAll(resp.Body)
+					if err != nil {
+						errs <- fmt.Errorf("error collecting %s: %w", vt, errBadRequest)
+						return
+					}
+					respContent := string(b)
+					// taken from pprof.Profile https://github.com/golang/go/blob/release-branch.go1.20/src/net/http/pprof/pprof.go#L133
+					if strings.Contains(respContent, "profile duration exceeds server's WriteTimeout") {
+						errs <- fmt.Errorf("%w: %s", ErrProfileTooLong, respContent)
+					} else {
+						errs <- fmt.Errorf("error collecting %s: %w: %s", vt, errBadRequest, respContent)
+					}
+				} else {
+					errs <- fmt.Errorf("error collecting %s: %w", vt, errBadRequest)
+				}
+				return
+			}
+			// write to file
+			f, err := os.Create(filepath.Join(genDir, vt))
+			if err != nil {
+				errs <- fmt.Errorf("error creating file for %s: %w", vt, err)
+				return
+			}
+			wc := utils.NewDeferableWriteCloser(f)
+			defer wc.Close()
+
+			_, err = io.Copy(wc, resp.Body)
+			if err != nil {
+				errs <- fmt.Errorf("error writing to file for %s: %w", vt, err)
+				return
+			}
+			err = wc.Close()
+			if err != nil {
+				errs <- fmt.Errorf("error closing file for %s: %w", vt, err)
+				return
+			}
+		}(vt)
+	}
+	wgPprof.Wait()
+	close(errs)
+	// Atmost one err is emitted per vital.
+	cli.Logger.Infof("collected %d/%d profiles", len(vitals)-len(errs), len(vitals))
+	if len(errs) > 0 {
+		var merr error
+		for err := range errs {
+			merr = errors.Join(merr, err)
+		}
+		return cli.errorOut(fmt.Errorf("profile collection failed:\n%v", merr))
+	}
+	return nil
 }
