+++ core/main_test.go
@@ -91,10 +91,12 @@ func ExampleRun_admin() {
 	//    core.test admin command [command options] [arguments...]
 	//
 	// COMMANDS:
-	//    chpass  Change your API password remotely
-	//    login   Login to remote client by creating a session cookie
-	//    logout  Delete any local sessions
-	//    users   Create, edit permissions, or delete API users
+	//    chpass   Change your API password remotely
+	//    login    Login to remote client by creating a session cookie
+	//    logout   Delete any local sessions
+	//    profile  Collects profile metrics from the node.
+	//    status   Displays the health of various services running inside the node.
+	//    users    Create, edit permissions, or delete API users
 	//
 	// OPTIONS:
 	//    --help, -h  show help
@@ -447,8 +449,6 @@ func ExampleRun_node() {
 	// COMMANDS:
 	//    start, node, n            Run the Chainlink node
 	//    rebroadcast-transactions  Manually rebroadcast txs matching nonce range with the specified gas price. This is useful in emergencies e.g. high gas prices and/or network congestion to forcibly clear out the pending TX queue
-	//    status                    Displays the health of various services running inside the node.
-	//    profile                   Collects profile metrics from the node.
 	//    validate                  Validate the TOML configuration and secrets that are passed as flags to the `node` command. Prints the full effective configuration, with defaults included
 	//    db                        Commands for managing the database.
 	//
